+++ lib/ui/academics/AcademicsHomePanel.dart
@@ -14,19 +14,23 @@
  * limitations under the License.
  */
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/CheckList.dart';
 import 'package:illinois/service/Config.dart';
+import 'package:illinois/service/DeepLink.dart';
 import 'package:illinois/service/FlexUI.dart';
+import 'package:illinois/service/Guide.dart';
+import 'package:illinois/ui/WebPanel.dart';
 import 'package:illinois/ui/academics/AcademicsEventsContentWidget.dart';
 import 'package:illinois/ui/academics/SkillsSelfEvaluation.dart';
 import 'package:illinois/ui/academics/StudentCourses.dart';
 import 'package:illinois/ui/canvas/CanvasCoursesContentWidget.dart';
 import 'package:illinois/ui/gies/CheckListContentWidget.dart';
-import 'package:illinois/utils/AppUtils.dart';
-import 'package:rokwire_plugin/service/connectivity.dart';
+import 'package:illinois/ui/guide/GuideDetailPanel.dart';
+import 'package:illinois/ui/wellness/todo/WellnessToDoHomeContentWidget.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/widgets/TabBar.dart' as uiuc;
@@ -54,6 +58,7 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
   late AcademicsContent _selectedContent;
   List<AcademicsContent>? _contentValues;
   bool _contentValuesVisible = false;
+  UniqueKey _dueDateCatalogKey = UniqueKey();
 
   @override
   void initState() {
@@ -229,12 +234,14 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
       return AcademicsContent.student_courses;
     } else if (code == 'academics_events') {
       return AcademicsContent.events;
+    } else if (code == 'skills_self_evaluation') {
+      return AcademicsContent.skills_self_evaluation;
+    } else if (code == 'todo_list') {
+      return AcademicsContent.todo_list;
     } else if (code == 'due_date_catalog') {
       return AcademicsContent.due_date_catalog;
     } else if (code == 'my_illini') {
       return AcademicsContent.my_illini;
-    } else if (code == 'skills_self_evaluation') {
-      return AcademicsContent.skills_self_evaluation;
     } else {
       return null;
     }
@@ -242,15 +249,25 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
 
   void _onTapContentItem(AcademicsContent contentItem) {
     Analytics().logSelect(target: '$contentItem');
+    String? launchUrl;
     if (contentItem == AcademicsContent.my_illini) {
       // Open My Illini in an external browser
-      _onMyIlliniSelected();
+      //_onMyIlliniSelected();
+      launchUrl = Config().myIlliniUrl;
     } else if (contentItem == AcademicsContent.due_date_catalog) {
       // Open Due Date Catalog in an external browser
-      _onTapDueDateCatalog();
+      launchUrl = Config().dateCatalogUrl;
     } else {
       _selectedContent = _lastSelectedContent = contentItem;
     }
+
+    if ((launchUrl != null) && (Guide().detailIdFromUrl(launchUrl) == null)) {
+      _launchUrl(launchUrl);
+    }
+    else {
+      _selectedContent = _lastSelectedContent = contentItem;
+    }
+
     _changeSettingsContentValuesVisibility();
   }
 
@@ -266,7 +283,7 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
     }
   }
 
-  void _onMyIlliniSelected() {
+  /*void _onMyIlliniSelected() {
     if (Connectivity().isOffline) {
       AppAlert.showOfflineMessage(context,
           Localization().getStringEx('panel.browse.label.offline.my_illini', 'My Illini not available while offline.'));
@@ -294,14 +311,31 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
       //   Navigator.push(context, CupertinoPageRoute(builder: (context) => WebPanel(url: Config().myIlliniUrl, title: myIlliniPanelTitle,)));
       // }
     }
-  }
+  }*/
 
-  void _onTapDueDateCatalog() {
+  /*void _onTapDueDateCatalog() {
     Analytics().logSelect(target: "Due Date Catalog");
     if (StringUtils.isNotEmpty(Config().dateCatalogUrl)) {
       Uri? uri = Uri.tryParse(Config().dateCatalogUrl!);
       if (uri != null) {
         launchUrl(uri);
+      }
+    }
+  }*/
+
+  void _launchUrl(String? url, { bool launchInternal = false}) {
+    if (StringUtils.isNotEmpty(url)) {
+      if (DeepLink().isAppUrl(url)) {
+        DeepLink().launchUrl(url);
+      }
+      else if (launchInternal && UrlUtils.launchInternal(url)){
+        Navigator.push(context, CupertinoPageRoute(builder: (context) => WebPanel(url: url)));
+      }
+      else {
+        Uri? uri = Uri.tryParse(url!);
+        if (uri != null) {
+          launchUrl(uri);
+        }
       }
     }
   }
@@ -317,7 +351,7 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
   }
 
   Widget get _rawContentWidget {
-    // There is no content for AcademicsContent.my_illini and AcademicsContent.due_date_catalog - it is a web url opened in an external browser
+    // There is no content for AcademicsContent.my_illini - it is a web url opened in an external browser
     switch (_selectedContent) {
       case AcademicsContent.events:
         return AcademicsEventsContentWidget();
@@ -331,6 +365,11 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
         return StudentCoursesContentWidget();
       case AcademicsContent.skills_self_evaluation:
         return SkillsSelfEvaluation();
+      case AcademicsContent.todo_list:
+        return WellnessToDoHomeContentWidget();
+      case AcademicsContent.due_date_catalog:
+        String? guideId = Guide().detailIdFromUrl(Config().dateCatalogUrl);
+        return (guideId != null) ? GuideDetailWidget(key: _dueDateCatalogKey, guideEntryId: guideId, headingColor: Styles().colors?.background) : Container();
       default:
         return Container();
     }
@@ -358,12 +397,14 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
         return Localization().getStringEx('panel.academics.section.canvas_courses.label', 'My Gies Canvas Courses');
       case AcademicsContent.student_courses:
         return Localization().getStringEx('panel.academics.section.student_courses.label', 'My Courses');
+      case AcademicsContent.skills_self_evaluation:
+        return Localization().getStringEx('panel.academics.section.skills_self_evaluation.label', 'Skills Self-Evaluation');
+      case AcademicsContent.todo_list:
+        return Localization().getStringEx('panel.academics.section.todo_list.label', 'To-Do List');
       case AcademicsContent.due_date_catalog:
         return Localization().getStringEx('panel.academics.section.due_date_catalog.label', 'Due Date Catalog');
       case AcademicsContent.my_illini:
         return Localization().getStringEx('panel.academics.section.my_illini.label', 'myIllini');
-      case AcademicsContent.skills_self_evaluation:
-        return Localization().getStringEx('panel.academics.section.skills_self_evaluation.label', 'Skills Self-Evaluation');
     }
   }
 
@@ -379,4 +420,4 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
   }
 }
 
-enum AcademicsContent { events, gies_checklist, uiuc_checklist, canvas_courses, student_courses, due_date_catalog, my_illini, skills_self_evaluation }
+enum AcademicsContent { events, gies_checklist, uiuc_checklist, canvas_courses, student_courses, skills_self_evaluation, todo_list, due_date_catalog, my_illini }