+++ lib/ui/settings/SettingsMealPlanPanel.dart
@@ -483,7 +483,7 @@ class _SettingsMealPlanPanelState extends State<SettingsMealPlanPanel> implement
   Widget _buildBalanceTableHeaderItem(String text){
     return _buildBalanceTableItem(text: text, backColor: Styles().colors!.fillColorPrimaryVariant,
         showBorder: false,
-        textStyle: Styles().textStyles?.getTextStyle("widget.heading.medium"));
+        textStyle: Styles().textStyles?.getTextStyle("widget.heading.medium_small"));
   }
 
   Widget _buildBalanceTableItem({required String text, bool showBorder = true, Color? backColor, TextStyle? textStyle, TextAlign textAlign = TextAlign.left}) {