+++ lib/ui/settings/SettingsPersonalInfoContentWidget.dart
@@ -415,7 +415,7 @@ class _SettingsPersonalInfoContentWidgetState extends State<SettingsPersonalInfo
     return await AppAlert.showCustomDialog(context: context,
       contentWidget:
         Text(Localization().getStringEx('panel.settings.home.calendar.research.prompt.title', promptEn),
-          style: TextStyle(fontFamily: Styles().fontFamilies?.regular, fontSize: 16, color: Styles().colors?.fillColorPrimary,),
+          style: Styles().textStyles?.getTextStyle("widget.message.regular"),
         ),
       actions: [
         TextButton(
@@ -547,7 +547,7 @@ class _SettingsPersonalInfoContentWidgetState extends State<SettingsPersonalInfo
               child: Text(
                 Localization().getStringEx("panel.profile_info.logout.message", promptEn),
                 textAlign: TextAlign.left,
-                style: Styles().textStyles?.getTextStyle("widget.dialog.message.dark.medium")
+                style: Styles().textStyles?.getTextStyle("widget.dialog.message.dark.regular")
               ),
             ),
             Row(
@@ -805,7 +805,7 @@ class _SettingsPersonalInfoContentWidgetState extends State<SettingsPersonalInfo
     AppAlert.showCustomDialog(context: context,
         contentWidget:
         Text(promptEn,
-          style: TextStyle(fontFamily: Styles().fontFamilies?.regular, fontSize: 16, color: Styles().colors?.fillColorPrimary,),
+          style: Styles().textStyles?.getTextStyle("widget.message.regular"),
         ),
         actions: [
           TextButton(