+++ lib/ui/settings/SettingsVerifyIdentityPanel.dart
@@ -70,7 +70,7 @@ class _SettingsVerifyIdentityPanelState extends State<SettingsVerifyIdentityPane
           Container(padding: EdgeInsets.symmetric(horizontal: 24),
               child:RichText(
                 text: TextSpan(
-                  style:  Styles().textStyles?.getTextStyle("widget.info.regular"),
+                  style:  Styles().textStyles?.getTextStyle("widget.info.regular.thin"),
                   children: <TextSpan>[
                     TextSpan(text: Localization().getStringEx("panel.settings.verify_identity.label.connect_id.desription1", "Are you a ")),
                     TextSpan(text: Localization().getStringEx("panel.settings.verify_identity.label.connect_id.desription2", "university student"),
@@ -94,7 +94,7 @@ class _SettingsVerifyIdentityPanelState extends State<SettingsVerifyIdentityPane
           Container(padding: EdgeInsets.symmetric(horizontal: 24),
               child:RichText(
                 text: TextSpan(
-                  style:  Styles().textStyles?.getTextStyle("widget.info.regular"),
+                  style:  Styles().textStyles?.getTextStyle("widget.info.regular.thin"),
                   children: <TextSpan>[
                     TextSpan(text: Localization().getStringEx("panel.settings.verify_identity.label.phone_or_email.desription1", "Don’t have a NetID"),
                         style:  Styles().textStyles?.getTextStyle("widget.info.regular.fat")),