+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [4.3.19] - 2023-02-23
+### Added
+- Switch of the data source for Canvas Courses [#3006](https://github.com/rokwire/illinois-app/issues/3006).
+
+## [4.3.18] - 2023-02-22
+### Changed
+- Updated default initial selection in Maps [#1942](https://github.com/rokwire/illinois-app/issues/1942).
+- Removed To-Do List from Wellness Section [#2951](https://github.com/rokwire/illinois-app/issues/2951).
+- Embed guide content in Academics and Wellness home panels [#3004](https://github.com/rokwire/illinois-app/issues/3004).
+
+## [4.3.17] - 2023-02-20
+- N/A
+
 ## [4.3.16] - 2023-02-14
 - N/A
 
@@ -92,7 +105,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Switch to xCode 14.2.
 
+## [4.2.53] - 2023-02-20
+### Changed
+- Update MTD text to include “Bus” [#2952](https://github.com/rokwire/illinois-app/issues/2952).
+- Move To-Do List to Academics Section [#2951](https://github.com/rokwire/illinois-app/issues/2951).
+
 ## [4.2.52] - 2023-02-16
+### Changed
 - Launch privacy policy web content in an external web browser on iOS platforms [#2909](https://github.com/rokwire/illinois-app/issues/2909).
 - Launch feedback web content in an external web browser on iOS platforms [#2909](https://github.com/rokwire/illinois-app/issues/2909).